+++ src/helpers/index.ts
@@ -1,0 +1,2 @@
+export { comparePasswords } from './comparePasswords.helper';
+export { setupSwagger } from './swaggerSetup.helper';
