+++ src/main.ts
@@ -4,7 +4,7 @@ import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { NestExpressApplication } from '@nestjs/platform-express';
-import { setupSwagger } from './utils';
+import { setupSwagger } from './helpers';
 import { setupSecurity } from './middlewares';
 
 config();
