+++ src/modules/users/users.service.ts
@@ -8,7 +8,7 @@ import { UserEntity } from '../../models/entities/user.entity';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { RegisterUserDto } from './dto/create-user.dto';
 import { LoginCredentialsPayload } from '../auth/interface/payload.interface';
-import { comparePasswords } from '../../utils';
+import { comparePasswords } from '../../helpers';
 import { ERRORS, PostgresErrorCode } from '../../constants';
 
 @Injectable()
