--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export { comparePasswords } from './comparePasswords';
-export { setupSwagger } from './swaggerSetup';
